# Enable OTel by default