--- a/cmd/config_validation_test.go
+++ b/cmd/config_validation_test.go
@@ -807,6 +807,7 @@ func TestValidateConfigFile_MetricsConfigSuccessful(t *testing.T) {
 				StackdriverExportInterval:      0,
 				CloudMetricsExportIntervalSecs: 0,
 				PrometheusPort:                 0,
+				EnableOtel:                     true,
 			},
 		},
 		{
@@ -814,6 +815,7 @@ func TestValidateConfigFile_MetricsConfigSuccessful(t *testing.T) {
 			configFile: "testdata/valid_config.yaml",
 			expectedConfig: &cfg.MetricsConfig{
 				CloudMetricsExportIntervalSecs: 10,
+				EnableOtel:                     true,
 			},
 		},
 	}