### Description
Use OTel for metrics by default. One can override the behavior to use OpenCensus instead by setting --enable-otel flag to false.

Perf:

| Branch | File Size  |   Read BW    |  Write BW  | RandRead BW  | RandWrite BW |
|--------|------------|--------------|------------|--------------|--------------|
| Master |  0.25MiB   | 480.23MiB/s  | 1.28MiB/s  |  67.36MiB/s  |  1.24MiB/s   |
|   PR   |  0.25MiB   | 489.49MiB/s  | 1.21MiB/s  |  68.74MiB/s  |  1.32MiB/s   |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 48.828MiB  | 3535.15MiB/s | 81.58MiB/s | 1250.43MiB/s |  82.85MiB/s  |
|   PR   | 48.828MiB  | 3428.55MiB/s | 81.17MiB/s | 1231.22MiB/s |  81.53MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 976.562MiB | 3420.17MiB/s | 30.88MiB/s | 412.77MiB/s  |  33.81MiB/s  |
|   PR   | 976.562MiB | 3489.39MiB/s | 33.64MiB/s | 482.15MiB/s  |  32.29MiB/s  |

### Link to the issue in case of a bug fix.
b/387898221

### Testing details
1. Manual - yes
2. Unit tests - NA
3. Integration tests - Already tested vi existing CI tests